Add full URL and https support for the Gluetun API
