+++ pkg/linkstats/types/linkstats.go
